+++ meerket/meerket-api/src/main/java/org/j1p5/api/auth/service/OauthClientService.java
@@ -1,19 +1,21 @@
-package org.j1p5.domain.auth;
+package org.j1p5.api.auth.service;
 
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
+
+import org.j1p5.domain.auth.OauthClient;
 import org.j1p5.domain.auth.dto.OauthProfile;
 import org.j1p5.domain.auth.dto.OauthToken;
 import org.j1p5.domain.user.entity.Provider;
 import org.springframework.stereotype.Component;
 
 @Component
-public class OauthSender {
+public class OauthClientService {
 
     private final Map<Provider, OauthClient> clients;
 
-    public OauthSender(List<OauthClient> oauthClients) {
+    public OauthClientService(List<OauthClient> oauthClients) {
         this.clients =
                 oauthClients.stream()
                         .collect(
@@ -21,8 +23,8 @@ public class OauthSender {
                                         OauthClient::getProvider, oauthClient -> oauthClient));
     }
 
-    public OauthProfile request(String code, String provider) {
-        OauthClient oauthClient = clients.get(Provider.get(provider));
+    public OauthProfile request(String code, Provider provider) {
+        OauthClient oauthClient = clients.get(provider);
 
         OauthToken token = oauthClient.getOauthToken(code);
